+++ test/hexapdf/type/acro_form/test_variable_text_field.rb
@@ -65,19 +65,24 @@ describe HexaPDF::Type::AcroForm::VariableTextField do
   end
 
   describe "parse_default_appearance_string" do
+    before do
+      @color = HexaPDF::Content::ColorSpace.prenormalized_device_color([1])
+    end
+
     it "parses the default appearance string of the field" do
       @field[:DA] = "1 g //F1 20 Tf 5 w /F2 10 Tf"
-      assert_equal([:F2, 10], @field.parse_default_appearance_string)
+      assert_equal([:F2, 10, @color], @field.parse_default_appearance_string)
     end
 
     it "uses the default appearance string of a parent field" do
       parent = @doc.add({DA: "/F1 15 Tf"}, type: :XXAcroFormField)
       @field[:Parent] = parent
-      assert_equal([:F1, 15], @field.parse_default_appearance_string)
+      assert_equal([:F1, 15, nil], @field.parse_default_appearance_string)
     end
 
     it "uses the global default appearance string" do
-      assert_equal([:F1, 0], @field.parse_default_appearance_string)
+      assert_equal([:F1, 0, HexaPDF::Content::ColorSpace.prenormalized_device_color([0])],
+                   @field.parse_default_appearance_string)
     end
 
     it "fails if no /DA value is set" do